career_mentor_agent/pyproject.toml
[project]
name = "career-mentor-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chainlit>=2.6.7",
    "openai-agents>=0.2.3",
    "python-dotenv>=1.1.1",
]
